Revamping data aggregation.

+++ src/util/FloatArray.java
@@ -1,39 +1,39 @@
 package util;
 
 /**
- * Object that stores an array of ints in a contiguous region of an int[]. Meant for quick and efficient data
+ * Object that stores an array of ints in a contiguous region of an float[]. Meant for quick and efficient data
  * aggregation. Once this is constructed from an array, the array should be considered invalid and never used again.
  * Once toArray is called to extract the data, this object should not be used again as the actions likely will not have
  * effect on the array.
  */
-public abstract class IntArray {
+public abstract class FloatArray {
 
 	protected static final int	DEFAULT_CAPACITY	= 8;
 
-	protected int[]				data;
+	protected float[]			data;
 	protected int				size;
 
-	public IntArray() {
-		data = new int[DEFAULT_CAPACITY];
+	public FloatArray() {
+		data = new float[DEFAULT_CAPACITY];
 	}
 
 	/**
-	 * Creates a new IntArray backed by the provided array. Modifying the array may render this IntArray invalid and
+	 * Creates a new FloatArray backed by the provided array. Modifying the array may render this FloatArray invalid and
 	 * vice-versa. Thus the provided array should no longer be modified after being provided to this constructor.
 	 */
 
-	protected IntArray(int... data) {
+	protected FloatArray(float... data) {
 		this.data = data;
 		size = data.length;
 	}
 
-	public abstract void append(int i);
+	public abstract void append(float i);
 
-	public abstract void prepend(int i);
+	public abstract void prepend(float i);
 
-	public abstract void append(int... ints);
+	public abstract void append(float... ints);
 
-	public abstract void prepend(int... ints);
+	public abstract void prepend(float... ints);
 
 	/**
 	 * Gets the index of the first piece of data. The data will be in the range from [getDataStart(), getDataStart() +
@@ -42,28 +42,28 @@ public abstract class IntArray {
 	protected abstract int getDataStart();
 
 	/**
-	 * Appends the IntArray provided to the end of this IntArray.
+	 * Appends the FloatArray provided to the end of this IntArray.
 	 */
-	public abstract void append(IntArray array);
+	public abstract void append(FloatArray array);
 
 	/**
-	 * Appends the IntArray provided to the beginning of this IntArray.
+	 * Appends the FloatArray provided to the beginning of this IntArray.
 	 */
-	public abstract void prepend(IntArray array);
+	public abstract void prepend(FloatArray array);
 
 	/**
 	 * Appends the IntArrays provided to the end of this IntArray.
 	 */
-	public abstract void append(IntArray... arrays);
+	public abstract void append(FloatArray... arrays);
 
 	/**
 	 * Appends the IntArrays provided to the beginning of this IntArray.
 	 */
-	public abstract void prepend(IntArray... arrays);
+	public abstract void prepend(FloatArray... arrays);
 
-	public abstract int removeFirst();
+	public abstract float removeFirst();
 
-	public abstract int removeLast();
+	public abstract float removeLast();
 
 	/**
 	 * Returns the number of elements held in this IntArray.
@@ -78,9 +78,9 @@ public abstract class IntArray {
 	public abstract void trim();
 
 	/**
-	 * Trim and returns the backing array. Modifying the array may render this IntArray invalid and vice-versa.
+	 * Trim and returns the backing array. Modifying the array may render this FloatArray invalid and vice-versa.
 	 */
-	public abstract int[] toArray();
+	public abstract float[] toArray();
 
 	/**
 	 * Grows the backing array so that it can hold at least min elements. Does nothing if the backing array is already
@@ -89,12 +89,12 @@ public abstract class IntArray {
 	protected abstract void growTo(int min);
 
 	/**
-	 * Returns a new int[] capable of holding at least min elements. Returns a new array of length size if min < size.
+	 * Returns a new float[] capable of holding at least min elements. Returns a new array of length size if min < size.
 	 */
-	protected int[] newArray(int min) {
+	protected float[] newArray(int min) {
 		if (shouldGrow(min))
-			return new int[growToSize(min)];
-		return new int[data.length];
+			return new float[growToSize(min)];
+		return new float[data.length];
 	}
 
 	protected boolean shouldGrow(int min) {
